pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llmbase"
version = "1.6.0"
description = "Unified API for Multiple LLM Providers - OpenAI, Anthropic, Gemini, DeepSeek, Perplexity, Ollama"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Ng Chong", email = "your.email@example.com"}
]
maintainers = [
    {name = "Ng Chong", email = "your.email@example.com"}
]
keywords = [
    "llm",
    "ai",
    "llmbase",
    "openai",
    "anthropic",
    "claude",
    "gemini",
    "deepseek",
    "ollama",
    "api",
    "wrapper",
    "unified",
    "agentic",
    "rag"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "google-genai>=1.0.0",
    "requests>=2.31.0",
    "urllib3>=2.0.0",
]

[project.optional-dependencies]
# Core dependencies (Library mode - no Flask)
core = [
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "google-genai>=1.0.0",
    "requests>=2.31.0",
    "urllib3>=2.0.0",
]
# API server mode (with Flask)
api = [
    "llmservices[core]",
    "flask>=3.0.0",
    "flask-session>=0.5.0",
]
# All optional dependencies
all = [
    "llmservices[api]",
]
# Development dependencies
dev = [
    "llmservices[all]",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/llmbase"
Documentation = "https://github.com/yourusername/llmbase#readme"
Repository = "https://github.com/yourusername/llmbase"
Issues = "https://github.com/yourusername/llmbase/issues"

[tool.setuptools]
py-modules = ["llmservices"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.txt", "*.md", "*.html"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312', 'py313']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "flask.*",
    "flask_session.*",
    "google.*",
    "genai.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=llmservices --cov-report=term-missing"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["llmservices"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
]
